move info container types into subfolders, improve pathing for testing
- i moved the info container types because i decided i actually dont like any files other than the startup class being at the top level of a project, and it also makes their purposes more clear.
- i added different debug and release base directories for convenience of locating test output files.
- i added a console message showing the log file path for convenience in finding it after the tool runs.
