+++ source/App/Program.cs
@@ -1,5 +1,5 @@
 ﻿using Petrichor.App.Utilities;
-using Petrichor.Common;
+using Petrichor.Common.Info;
 using Petrichor.Logging;
