+++ source/TestShared/Info/TestDirectories.cs
@@ -1,0 +1,16 @@
+using Petrichor.Common.Info;
+
+
+namespace Petrichor.TestShared.Info
+{
+	public class TestDirectories : ProjectDirectories
+	{
+		private const string testDirectoryName = "_test";
+
+
+		public static string TestDirectory => $@"{BaseDirectory}..\..\..\..\{testDirectoryName}";
+		public static string TestInputDirectory => $@"{TestDirectory}\{inputDirectoryName}";
+		public static string TestLogDirectory => $@"{TestDirectory}\{logDirectoryName}";
+		public static string TestOutputDirectory => $@"{TestDirectory}\{outputDirectoryName}";
+	}
+}
