+++ source/TestShared/Utilities/TestUtilities.cs
@@ -1,4 +1,5 @@
 ﻿using Petrichor.Logging;
+using Petrichor.TestShared.Info;
 
 
 namespace Petrichor.TestShared.Utilities
